requirements.txt
fake-useragent~=1.1.3
aiogram~=3.1.1
python-dotenv~=1.0.0
loguru~=0.7.2
requests~=2.31.0
beautifulsoup4~=4.12.2
rss-parser~=0.2.4
SQLAlchemy~=2.0.22